+++ src/Import/JsonImport.php
@@ -1,6 +1,6 @@
 <?php
 /**
- * CsvImport class
+ * JsonImport class
  *
  * @license    GPL v2 or later (http://www.gnu.org/licenses/gpl.html)
  * @author     Didier Corbière <contact@atoll-digital-library.org>
@@ -9,10 +9,6 @@
 
 namespace Marsender\EPubLoader\Import;
 
-use Marsender\EPubLoader\Metadata\BookInfos;
-use Marsender\EPubLoader\Metadata\GoodReads\BookShowResult;
-use Marsender\EPubLoader\Metadata\GoogleBooks\SearchResult;
-use Marsender\EPubLoader\Metadata\GoogleBooks\Volume;
 use Marsender\EPubLoader\RequestHandler;
 use Exception;
 
@@ -32,11 +28,15 @@ class JsonImport extends BookImport
         $nbOk = 0;
         $nbError = 0;
         if (!empty($data["kind"]) && $data["kind"] == "books#volumes") {
-            $result = SearchResult::fromJson($data);
+            // Parse the JSON data
+            $result = GoogleBooksVolume::parseResult($data);
+            if (empty($result->getItems())) {
+                $result->items = [];
+            }
             foreach ($result->getItems() as $volume) {
                 try {
                     // Load the book infos
-                    $bookInfos = self::loadFromGoogleBooksVolume($inBasePath, $volume);
+                    $bookInfos = GoogleBooksVolume::load($inBasePath, $volume);
                     // Add the book
                     $this->addBook($bookInfos, 0);
                     $nbOk++;
@@ -47,23 +47,25 @@ class JsonImport extends BookImport
                 }
             }
         } elseif (!empty($data["kind"]) && $data["kind"] == "books#volume") {
-            $volume = Volume::fromJson($data);
             try {
+                // Parse the JSON data
+                $volume = GoogleBooksVolume::parse($data);
                 // Load the book infos
-                $bookInfos = self::loadFromGoogleBooksVolume($inBasePath, $volume);
+                $bookInfos = GoogleBooksVolume::load($inBasePath, $volume);
                 // Add the book
                 $this->addBook($bookInfos, 0);
                 $nbOk++;
             } catch (Exception $e) {
-                $id = $volume->getId() ?? spl_object_hash($volume);
+                $id = basename($fileName);
                 $errors[$id] = $e->getMessage();
                 $nbError++;
             }
         } elseif (!empty($data["page"]) && $data["page"] == "/book/show/[book_id]") {
-            $bookShowResult = BookShowResult::fromJson($data);
             try {
+                // Parse the JSON data
+                $bookShowResult = GoodReadsBook::parse($data);
                 // Load the book infos
-                $bookInfos = self::loadFromGoodReadsBook($inBasePath, $bookShowResult);
+                $bookInfos = GoodReadsBook::load($inBasePath, $bookShowResult);
                 // Add the book
                 $this->addBook($bookInfos, 0);
                 $nbOk++;
@@ -77,116 +79,6 @@ class JsonImport extends BookImport
         }
         $message = sprintf('Import ebooks from %s - %d files OK - %d files Error', $fileName, $nbOk, $nbError);
         return [$message, $errors];
-    }
-
-    /**
-     * Loads book infos from an Google Books volume
-     *
-     * @param string $inBasePath base directory
-     * @param Volume $volume Google Books volume
-     * @throws Exception if error
-     *
-     * @return BookInfos
-     */
-    public static function loadFromGoogleBooksVolume($inBasePath, $volume)
-    {
-        $volumeInfo = $volume->getVolumeInfo();
-        if (empty($volumeInfo)) {
-            throw new Exception('Invalid format for Google Books Volume');
-        }
-
-        $bookInfos = new BookInfos();
-        $bookInfos->mBasePath = $inBasePath;
-        // @todo check accessInfo for epub, pdf etc.
-        $bookInfos->mFormat = 'epub';
-        // @todo use calibre_external_storage in COPS
-        $bookInfos->mPath = (string) $volume->getSelfLink();
-        if (str_starts_with($bookInfos->mPath, $inBasePath)) {
-            $bookInfos->mPath = substr($bookInfos->mPath, strlen($inBasePath) + 1);
-        }
-        $bookInfos->mName = (string) $volume->getId();
-        $bookInfos->mUuid = 'google:' . $bookInfos->mName;
-        $bookInfos->mUri = (string) ($volumeInfo->getCanonicalVolumeLink() ?? $volume->getSelfLink());
-        $bookInfos->mTitle = (string) $volumeInfo->getTitle();
-        $authors = [];
-        foreach ($volumeInfo->getAuthors() as $author) {
-            $authorSort = BookInfos::getSortString($author);
-            $authors[$authorSort] = $author;
-        }
-        $bookInfos->mAuthors = $authors;
-        $bookInfos->mLanguage = (string) $volumeInfo->getLanguage();
-        $bookInfos->mDescription = (string) $volumeInfo->getDescription();
-        $bookInfos->mSubjects = $volumeInfo->getCategories();
-        $bookInfos->mCover = (string) $volumeInfo->getImageLinks()?->getThumbnail();
-        $identifiers = $volumeInfo->getIndustryIdentifiers();
-        if (!empty($identifiers)) {
-            foreach ($identifiers as $identifier) {
-                if ($identifier->getType() == 'ISBN_13') {
-                    $bookInfos->mIsbn = $identifier->getIdentifier();
-                    break;
-                }
-                if ($identifier->getType() == 'ISBN_10') {
-                    $bookInfos->mIsbn = $identifier->getIdentifier();
-                    break;
-                }
-            }
-        }
-        //$bookInfos->mRights = $inArray[$i++];
-        $bookInfos->mPublisher = (string) $volumeInfo->getPublisher();
-        $series = $volumeInfo->getSeriesInfo();
-        if (!empty($series)) {
-            $bookInfos->mSerieIndex = (string) $series->getBookDisplayNumber();
-            // @todo use title to get series name
-            if (str_contains($bookInfos->mTitle, ':')) {
-                [$seriesName, $title] = explode(':', $bookInfos->mTitle, 2);
-                $seriesName = preg_replace('/\s*Vol.\s*/', '', preg_replace('/\s*\d+\s*/', '', $seriesName));
-                $bookInfos->mSerie = trim($seriesName);
-            } elseif (!empty($series->getVolumeSeries())) {
-                $info = $series->getVolumeSeries()[0];
-                // @todo get series name from id
-                $bookInfos->mSerie = (string) $info->getSeriesId();
-            }
-        }
-        $bookInfos->mCreationDate = (string) $volumeInfo->getPublishedDate();
-        // @todo no modification date here
-        $bookInfos->mModificationDate = $bookInfos->mCreationDate;
-        // Timestamp is used to get latest ebooks
-        $bookInfos->mTimeStamp = $bookInfos->mCreationDate;
-
-        return $bookInfos;
-    }
-
-    /**
-     * Loads book infos from an GoodReads book
-     *
-     * @param string $inBasePath base directory
-     * @param BookShowResult $bookShowResult GoodReads book show
-     * @throws Exception if error
-     *
-     * @return BookInfos
-     */
-    public static function loadFromGoodReadsBook($inBasePath, $bookShowResult)
-    {
-        $state = $bookShowResult->getProps()?->getPageProps()?->getApolloState();
-        if (empty($state)) {
-            throw new Exception('Invalid state for GoodReads book');
-        }
-        $bookRef = $state->getRootQuery()?->getGetBookByLegacyId()?->getRef();
-        $bookMap = $state->getBookMap();
-        if (empty($bookRef) || empty($bookMap) || empty($bookMap[$bookRef])) {
-            throw new Exception('Invalid bookRef for GoodReads book');
-        }
-        $workRef = $bookMap[$bookRef]->getWork()?->getRef();
-        $workMap = $state->getWorkMap();
-        if (empty($workRef) || empty($workMap) || empty($workMap[$workRef])) {
-            throw new Exception('Invalid workRef for GoodReads book');
-        }
-
-        $bookInfos = new BookInfos();
-        // @todo ...
-        $bookInfos->createUuid();
-
-        return $bookInfos;
     }
 
     /**