+++ tests/GoogleBooksTest.php
@@ -115,16 +115,17 @@ class GoogleBooksTest extends TestCase
         $this->assertCount(0, $errors);
     }
 
-    protected function skipTestJsonImportPath(): void
+    public function testJsonImportPath(): void
     {
         $dbPath = dirname(__DIR__) . '/cache/google';
         $dbFile = $dbPath . '/metadata.db';
         $import = new JsonImport($dbFile, true);
 
-        $jsonPath = 'authors';
+        //$jsonPath = 'authors';
+        $jsonPath = 'titles';
         [$message, $errors] = $import->loadFromPath($dbPath, $jsonPath);
 
-        $expected = '/cache/google/authors/Arthur Conan Doyle.en.40.json - 40 files OK - 0 files Error';
+        $expected = '/cache/google/titles/Émile Zola.La curée.fr.json - 10 files OK - 0 files Error';
         $this->assertStringContainsString($expected, $message);
         $this->assertCount(0, $errors);
     }