+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 1.5.x - 2024xxxx Maintenance release for 1.x (PHP >= 7.4)
   * ...
 
+3.3.2 - 20241001 Import books from JSON files (GoodReads API)
+  * Add JsonImport for GoodReads API cached files
+
 3.3.1 - 20240930 Import books from CSV and JSON files (Google Books API)
   * Clarify license GPL version 2 or later + fix file headers
   * Add JsonImport for Google Books API cached files